Conflicts:
	src/openMVG/matching_image_collection/Cascade_Hashing_Matcher_Regions.cpp
	src/openMVG/matching_image_collection/Cascade_Hashing_Matcher_Regions.hpp
	src/openMVG/matching_image_collection/Matcher.hpp
	src/openMVG/matching_image_collection/Matcher_Regions.cpp
	src/openMVG/matching_image_collection/Matcher_Regions.hpp
	src/openMVG/sfm/pipelines/sfm_regions_provider.hpp
	src/software/SfM/main_ComputeMatches.cpp

+++ src/openMVG/sfm/sfm_data_triangulation.cpp
@@ -4,10 +4,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "openMVG/sfm/sfm_data_triangulation.hpp"
-
 #include "openMVG/multiview/triangulation_nview.hpp"
 #include "openMVG/robust_estimation/rand_sampling.hpp"
+#include "openMVG/sfm/sfm_data.hpp"
+#include "openMVG/sfm/sfm_data_triangulation.hpp"
+
 #include "third_party/progress/progress.hpp"
 
 #include <deque>
@@ -225,7 +226,7 @@ const
   // - Ransac loop
   for (IndexT i = 0; i < nbIter; ++i)
   {
-    std::vector<size_t> vec_samples;
+    std::vector<uint32_t> vec_samples;
     robust::UniformSample(min_sample_index, obs.size(), &vec_samples);
     const std::set<IndexT> samples(vec_samples.begin(), vec_samples.end());
